+++ samples/Reading/Program.cs
@@ -54,7 +54,7 @@ internal static class Program
         {
             await foreach (var message in reader.Messages(cancellationToken))
             {
-                Console.WriteLine("Received: " + message.Value());
+                Console.WriteLine($"Received: {message.Value()}");
             }
         }
         catch (OperationCanceledException) { }