+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2022-02-02
+
+### Added
+
+- Support for on the fly authentication (AuthChallenge)
+- The IAuthentication interface for implementing custom authentication
+- IPulsarClientBuilder.Authentication(IAuthentication authentication) for using custom authentication
+- The AuthenticationFactory with support for:
+    - Token(string token)
+	- Token(Func\<CancellationToken, ValueTask\<string\>\> tokenSupplier)
+
+### Deprecated
+
+- IPulsarClientBuilder.AuthenticateUsingToken(string token) is marked as obsolete. Use Authentication(AuthenticationFactory.Token(...)) instead
+
 ## [2.0.1] - 2021-11-12
 
 ### Changed